Changed icon wallet to the connected digital wallet
